backend/backend-api/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.1
gunicorn>=21.2.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
aiofiles>=23.2.1

authlib>=1.3.0
python-jose>=3.3.0
PyJWT>=2.8.0
passlib==1.7.4
bcrypt==4.1.2
itsdangerous>=2.1.2
email-validator>=2.1.0
pytz>=2024.1

pymongo>=4.6.1
motor>=3.4.0
certifi>=2024.2.2
dnspython>=2.6.1
pydantic-settings>=2.2.1

APScheduler>=3.10.0
httpx>=0.27.0
cloudinary>=1.39.1

python-socketio>=5.11.0

structlog>=23.3.0
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=1.40.6

prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=6.1.0
psutil>=5.9.8
geopy>=2.4.1
redis[hiredis]>=5.0.0

pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
mongomock>=4.1.2
slowapi>=0.1.9
secure>=1.0.0

reportlab>=4.0,<5.0
webauthn>=2.0.0
python-magic>=0.4.27
Pillow>=10.0.0
